Moved editor templates to editor folder
