+++ lib/assets/javascripts/cartodb3/editor.js
@@ -25,7 +25,7 @@ var WidgetDefinitionsCollection = require('./data/widget-definitions-collection'
 var VisDefinitionModel = require('./data/vis-definition-model');
 var createEditorMenuTabPane = require('./components/tab-pane/create-editor-menu-tab-pane');
 var EditorPaneTemplate = require('./editor-pane.tpl');
-var EditorPaneIconItemTemplate = require('./editor-pane-icon.tpl');
+var EditorPaneIconItemTemplate = require('./editor/editor-pane-icon.tpl');
 var ModalsServiceModel = require('./components/modals/modals-service-model');
 var viewFactory = require('./components/view-factory');
 var UserModel = require('./data/user-model');
