+++ lib/assets/javascripts/cartodb3/editor/editor-pane-icon.tpl
