+++ lib/assets/javascripts/cartodb3/editor/editor-view.js
@@ -10,7 +10,7 @@ var CreatePrivacyOptions = require('../components/modals/privacy/create-privacy-
 var PrivacyCollection = require('../components/modals/privacy/privacy-collection');
 
 var Header = require('./editor-header.js');
-var EditorTabPaneTemplate = require('./editor-tab-pane.tpl');
+var EditorTabPaneTemplate = require('./editor/editor-tab-pane.tpl');
 var EditorWidgetsView = require('./widgets/widgets-view');
 var LayersView = require('./layers/layers-view');
 var ScrollView = require('../components/scroll/scroll-view');
